Use Context and extract translation zones component
